Initial commit - basic functionality implemented
Implemented ability to remove leading numbers within music file names.
For example, should a file be called "003 Oblivion.mp3", it would be
renamed as "Oblivion.mp3".

Implemented ability to scan over files and accumulate a database of
information about them, in regards to artists, albums and titles. Labels
cannot be duplicated, and they are displayed in a systemic fashion to
allow for easy reading.