Implemented adding music and code organisation
Added the ability to take a folder of tracks, scan over them, and move
them to the correct location within the playlist directory. The tracks
will be placed in their correct locations; in the correct artist and
album folders.

Reorganised a lot of the code to be more split up, whilst also moving
some features to the utilities.py file.